// client/src/app/shared/models/shopParams.ts
export class ShopParams {
  brands: string[] = [];
  types: string[] = [];
  sort = 'name';
  pageNumber = 1;
  pageSize = 10;
  search = '';
}


// export class ShopParams {
//   brands: string[] = [];
//   types: string[] = [];
//   sort = 'name';
//   pageNumber = 1;
//   pageSize = 10;
//   search = '';
// }